refactor: alert confirm으로 수정 - 프로젝트 삭제 취소 가능하도록 confirm창으로 수정

+++ src/pages/ProjectSetting/ProjectSettingContainer.tsx
@@ -62,7 +62,8 @@ const ProjectSettingContainer: FC = () => {
   };
 
   const handleDeleteButton = (projectIdToDelete: string) => {
-    alert('정말로 삭제하시겠습니까??');
+    const checkDelete = confirm('정말로 삭제하시겠습니까??');
+    if (!checkDelete) return;
     deleteProject(projectIdToDelete);
     dispatch(setCurrentProject(''));
     history.push('/project');
